Refactor: separação de arquivos auth e profile do service
