+++ src/service/Auth.ts
@@ -1,13 +1,5 @@
-import { UUID } from 'crypto';
-import {
-  UserForgot,
-  UserSignIn,
-  UserSingUp,
-  UserResponse,
-  MyProfileInfos,
-} from '../interface/UserInterface';
+import { UserForgot, UserSignIn, UserSingUp } from '../interface/UserInterface';
 import { apiAuth } from './api';
-import axios from 'axios';
 
 const signUp = async (data: UserSingUp) => {
   const response = await apiAuth.post(`/public/users/register`, data);
@@ -24,29 +16,4 @@ const forgot = async (data: UserForgot) => {
   return response;
 };
 
-const fetchUserData = async (id?: UUID): Promise<UserResponse> => {
-  if (!id) {
-    throw new Error('ID não foi encontrado');
-  }
-  const response = await apiAuth.get<UserResponse>(`/public/users/${id}`);
-
-  return response.data;
-};
-
-const fetchMyProfile = async (token?: string): Promise<MyProfileInfos> => {
-  if (!token) {
-    throw new Error('Usuário deve estar logado');
-  }
-
-  const api = axios.create({
-    baseURL: import.meta.env.VITE_API_URL,
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-  });
-  const response = await api.get<MyProfileInfos>(`/users/my-profile`);
-
-  return response.data;
-};
-
-export { signUp, signIn, forgot, fetchUserData, fetchMyProfile };
+export { signUp, signIn, forgot };
